Create Directory Links
Add functionality to directory items via useNavigate to allow
clickthrough to the relevant shop section
