+++ src/components/DirectoryItem/DirectoryItem.jsx
@@ -1,3 +1,5 @@
+import { useNavigate } from "react-router-dom";
+
 import {
   BackgroundImage,
   Body,
@@ -5,9 +7,13 @@ import {
 } from "./DirectoryItem.styles";
 
 const DirectoryItem = ({ category }) => {
-  const { title, imageUrl } = category;
+  const navigate = useNavigate();
+
+  const { title, imageUrl, route } = category;
+  const onNavigateHandler = () => navigate(route);
+
   return (
-    <DirectoryItemContainer>
+    <DirectoryItemContainer onClick={onNavigateHandler}>
       <BackgroundImage imageUrl={imageUrl} />
       <Body>
         <h2>{title}</h2>
